+++ lib/editor/widgets/flow_editor_grpc/flow_editor_grpc.dart
@@ -8,10 +8,10 @@ import 'package:trayce/common/config.dart';
 import 'package:trayce/editor/widgets/code_editor/code_editor_multi.dart';
 import 'package:trayce/editor/widgets/code_editor/code_editor_single.dart';
 import 'package:trayce/editor/widgets/common/form_table.dart';
-import 'package:trayce/editor/widgets/common/form_table_controller.dart';
 import 'package:trayce/editor/widgets/common/headers_table_read_only.dart';
 import 'package:trayce/editor/widgets/explorer/explorer_style.dart';
 import 'package:trayce/editor/widgets/flow_editor_http/focus_manager.dart';
+import 'package:trayce/editor/widgets/flow_editor_http/form_headers_controller.dart';
 
 import '../../../common/dropdown_style.dart';
 import '../../../common/style.dart';
@@ -61,7 +61,7 @@ class _FlowEditorGrpcState extends State<FlowEditorGrpc> with TickerProviderStat
   final CodeLineEditingController _responseController = CodeLineEditingController();
   final CodeLineEditingController _urlController = CodeLineEditingController();
   final CodeLineEditingController _bodyController = CodeLineEditingController();
-  late final FormTableController _headersController;
+  late final FormHeadersController _headersController;
   String _selectedMethod = 'GET';
   static const List<String> _httpMethods = ['GET', 'POST', 'PUT', 'DELETE', 'PATCH', 'HEAD', 'OPTIONS'];
 
@@ -77,8 +77,9 @@ class _FlowEditorGrpcState extends State<FlowEditorGrpc> with TickerProviderStat
 
     _focusManager = EditorFocusManager(context.read<EventBus>(), widget.tabKey);
 
-    _headersController = FormTableController(
+    _headersController = FormHeadersController(
       onStateChanged: () => setState(() {}),
+      initialRows: [],
       config: context.read<Config>(),
       focusManager: _focusManager,
       eventBus: context.read<EventBus>(),
